+++ src/core/form-layout/index.ts
@@ -1,5 +1,6 @@
 export * from './types';
 export * from './component';
+export * from './teleport-anchor';
 export * from './column';
 export * from './row';
 export * from './form-builder';