+++ src/core/index.ts
@@ -1,2 +1,3 @@
 export * as FormLayout from './form-layout';
-export { FormBuilder, FormBuilderBodyProp } from './form-layout';
+export { FormBuilder, FormBuilderBodyProp, useTeleportAnchor } from './form-layout';
+export type { TeleportAnchor } from './form-layout';